Check the dataset update access for "Add Resource"
Currently any user with the 'create dataset content' permission have access to
the "Add Resource" tab even if did not create the said dataset or his access
level should not allow to. This change will make the "Add Resource" visible to
only users with the 'edit own dataset content' or the 'edit any dataset content'
permissions.